requirements.txt
aiohttp==3.8.5
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.0.4
deepgram-sdk==0.2.4
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.74.1
frozenlist==1.3.0
h11==0.13.0
httptools==0.2.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
orjson
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML
requests==2.31.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
# uvloop==0.16.0
watchgod==0.7
websockets==10.2
yarl==1.7.2